Bug fixes + simplification
Fix: Factory was inefficient and crashing. Replaced the local dictionary with a class private dictionary with a different structure and used it to create object, using Activator.CreateInstance
Fix: Operation For crashed if one parameter for the loop was missing.
Change: Class Operation now owns the parameters (instead of the derived classes)
Add: Functions in Operation to access the operation description, a parameter, given the name, and all parameters because logic was repeated in derived classes.