NamedData as base class to Variable and Parameter (parcial)
Working to have Variable and Parameter sharing as much code as possible because most of the data that defines both is common (or seems to be).
Created namespace NamedData (better name, anyone?) and will create a class with the same name from which Variable and Parameter will derive and that will contain all enums and tables now in Variable.cs but outside of the class.
Still a lot to do. Hope it works.